Fixed: integer overflow for wav files > 2 GB
If a wav file is greater than 2 GB, m_iDataStart and m_iDataLen will both be negative and the wav file will fail to play.

+++ xbmc/cores/paplayer/WAVcodec.h
@@ -39,8 +39,8 @@ public:
 
 private:
   bool m_bHasFloat;
-  long m_iDataStart;
-  long m_iDataLen;
+  uint32_t m_iDataStart;
+  uint32_t m_iDataLen;
   DWORD m_ChannelMask;
 };
